# app/static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Finance Tracker API",
      "description": "RESTful API for tracking personal finances",
      "version": "1.0.0"
    },
    "basePath": "/api",
    "schemes": [
      "http"
    ],
    "securityDefinitions": {
      "Bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    },
    "paths": {
...
      "/transactions": {
        "get": {
          "summary": "Get all transactions",
          "tags": ["Transactions"],
          "security": [{"Bearer": []}],
          "produces": ["application/json"],
          "parameters": [
            {
              "in": "query",
              "name": "category_id",
              "type": "integer",
              "description": "Filter by category ID"
            },
            {
              "in": "query",
              "name": "type",
              "type": "string",
              "description": "Filter by transaction type (income or expense)"
            },
            {
              "in": "query",
              "name": "start_date",
              "type": "string",
              "description": "Filter by start date (YYYY-MM-DD)"
            },
            {
              "in": "query",
              "name": "end_date",
              "type": "string",
              "description": "Filter by end date (YYYY-MM-DD)"
            }
          ],
          "responses": {
            "200": {
              "description": "List of transactions"
            },
            "401": {
              "description": "Unauthorized"
            }
          }
        },
        "post": {
          "summary": "Create a new transaction",
          "tags": ["Transactions"],
          "security": [{"Bearer": []}],
          "consumes": ["application/json"],
          "produces": ["application/json"],
          "parameters": [
            {
              "in": "body",
              "name": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number"
                  },
                  "description": {
                    "type": "string"
                  },
                  "date": {
                    "type": "string",
                    "format": "date"
                  },
                  "type": {
                    "type": "string",
                    "enum": ["income", "expense"]
                  },
                  "category_id": {
                    "type": "integer"
                  }
                },
                "required": ["amount", "type", "category_id"]
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Transaction created successfully"
            },
            "400": {
              "description": "Invalid input"
            },
            "401": {
              "description": "Unauthorized"
            },
            "404": {
              "description": "Category not found"
            }
          }
        }
      },
      "/transactions/{id}": {
        "get": {
          "summary": "Get a specific transaction",
          "tags": ["Transactions"],
          "security": [{"Bearer": []}],
          "produces": ["application/json"],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "type": "integer",
              "description": "Transaction ID"
            }
          ],
          "responses": {
            "200": {
              "description": "Transaction details"
            },
            "401": {
              "description": "Unauthorized"
            },
            "404": {
              "description": "Transaction not found"
            }
          }
        },
        "put": {
          "summary": "Update a transaction",
          "tags": ["Transactions"],
          "security": [{"Bearer": []}],
          "consumes": ["application/json"],
          "produces": ["application/json"],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "type": "integer",
              "description": "Transaction ID"
            },
            {
              "in": "body",
              "name": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "amount": {
                    "type": "number"
                  },
                  "description": {
                    "type": "string"
                  },
                  "date": {
                    "type": "string",
                    "format": "date"
                  },
                  "type": {
                    "type": "string",
                    "enum": ["income", "expense"]
                  },
                  "category_id": {
                    "type": "integer"
                  }
                }
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Transaction updated successfully"
            },
            "400": {
              "description": "Invalid input"
            },
            "401": {
              "description": "Unauthorized"
            },
            "404": {
              "description": "Transaction or category not found"
            }
          }
        },
        "delete": {
          "summary": "Delete a transaction",
          "tags": ["Transactions"],
          "security": [{"Bearer": []}],
          "produces": ["application/json"],
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "type": "integer",
              "description": "Transaction ID"
            }
          ],
          "responses": {
            "200": {
              "description": "Transaction deleted successfully"
            },
            "401": {
              "description": "Unauthorized"
            },
            "404": {
              "description": "Transaction not found"
            }
          }
        }
      },
      "/transactions/summary": {
        "get": {
          "summary": "Get transaction summary",
          "tags": ["Transactions"],
          "security": [{"Bearer": []}],
          "produces": ["application/json"],
          "parameters": [
            {
              "in": "query",
              "name": "start_date",
              "type": "string",
              "description": "Start date for summary (YYYY-MM-DD)"
            },
            {
              "in": "query",
              "name": "end_date",
              "type": "string",
              "description": "End date for summary (YYYY-MM-DD)"
            }
          ],
          "responses": {
            "200": {
              "description": "Transaction summary"
            },
            "401": {
              "description": "Unauthorized"
            }
          }
        }
      },
...
    }
  }